refactor: move Windows-only cache types to cache_windows.go and document concurrency safety
