+++ receiver/windowseventlogreceiver/internal/sidcache/cache_windows.go
@@ -8,10 +8,13 @@ package sidcache
 import (
 	"errors"
 	"fmt"
+	"sync"
+	"sync/atomic"
 	"syscall"
 	"time"
 	"unsafe"
 
+	lru "github.com/hashicorp/golang-lru/v2"
 	"golang.org/x/sys/windows"
 )
 
@@ -81,7 +84,131 @@ type lsaTranslatedName struct {
 	DomainIndex int32
 }
 
-// lookupSID performs a Windows LSA lookup for the given SID string
+// cacheEntry wraps a ResolvedSID with TTL information
+type cacheEntry struct {
+	resolved  *ResolvedSID
+	expiresAt time.Time
+}
+
+// concurrentSIDCache implements the Cache interface with LRU eviction and TTL expiration.
+// It is safe for concurrent use by multiple goroutines.
+type concurrentSIDCache struct {
+	config Config
+	lru    *lru.Cache[string, *cacheEntry]
+
+	// Statistics (using atomic for lock-free thread safety)
+	hits      atomic.Uint64
+	misses    atomic.Uint64
+	evictions atomic.Uint64
+	errors    atomic.Uint64
+
+	mu sync.RWMutex
+}
+
+// New creates a new SID cache with the given configuration.
+// The returned Cache is safe for concurrent use.
+func New(config Config) (Cache, error) {
+	// Validate and set defaults
+	if config.Size == 0 {
+		config.Size = DefaultCacheSize
+	}
+	if config.TTL <= 0 {
+		config.TTL = DefaultCacheTTL
+	}
+
+	c := &concurrentSIDCache{
+		config: config,
+	}
+
+	// Create LRU cache with eviction callback
+	lruCache, err := lru.NewWithEvict(int(config.Size), func(_ string, _ *cacheEntry) {
+		c.evictions.Add(1)
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to create LRU cache: %w", err)
+	}
+
+	c.lru = lruCache
+	return c, nil
+}
+
+// Resolve looks up a SID and returns its resolved information.
+func (c *concurrentSIDCache) Resolve(sid string) (*ResolvedSID, error) {
+	// Validate SID format
+	if !isSIDFormat(sid) {
+		c.errors.Add(1)
+		return nil, fmt.Errorf("invalid SID format: %s", sid)
+	}
+
+	// Check well-known SIDs first (never cached, always available)
+	if resolved, ok := isWellKnownSID(sid); ok {
+		c.hits.Add(1)
+		return resolved, nil
+	}
+
+	// Check cache
+	c.mu.RLock()
+	entry, found := c.lru.Get(sid)
+	c.mu.RUnlock()
+
+	if found {
+		// Check if entry has expired
+		if time.Now().Before(entry.expiresAt) {
+			c.hits.Add(1)
+			return entry.resolved, nil
+		}
+		// Entry expired, remove it
+		c.mu.Lock()
+		c.lru.Remove(sid)
+		c.mu.Unlock()
+	}
+
+	// Cache miss - resolve via Windows API
+	c.misses.Add(1)
+	resolved, err := lookupSID(sid)
+	if err != nil {
+		c.errors.Add(1)
+		return nil, fmt.Errorf("failed to lookup SID %s: %w", sid, err)
+	}
+
+	// Add to cache with TTL
+	c.mu.Lock()
+	c.lru.Add(sid, &cacheEntry{
+		resolved:  resolved,
+		expiresAt: time.Now().Add(c.config.TTL),
+	})
+	c.mu.Unlock()
+
+	return resolved, nil
+}
+
+// Close releases resources held by the cache.
+// Purge is called explicitly rather than relying on GC to provide deterministic
+// cleanup of cached entries, which is useful for tests and follows the Cache
+// interface contract.
+func (c *concurrentSIDCache) Close() error {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	c.lru.Purge()
+	return nil
+}
+
+// Stats returns current cache statistics.
+func (c *concurrentSIDCache) Stats() Stats {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+
+	return Stats{
+		Hits:      c.hits.Load(),
+		Misses:    c.misses.Load(),
+		Evictions: c.evictions.Load(),
+		Size:      c.lru.Len(),
+		Errors:    c.errors.Load(),
+	}
+}
+
+// lookupSID performs a Windows LSA lookup for the given SID string.
 func lookupSID(sidString string) (*ResolvedSID, error) {
 	// Convert string SID to binary SID
 	var sid *windows.SID
@@ -186,7 +313,7 @@ func lookupSID(sidString string) (*ResolvedSID, error) {
 	}, nil
 }
 
-// sidTypeToString converts SID_NAME_USE to a readable string
+// sidTypeToString converts SID_NAME_USE to a readable string.
 func sidTypeToString(sidType uint32) string {
 	switch sidType {
 	case SidTypeUser:
